Changed plot function
The plots are now based on ggplot. They generally look a lot better now, but there are a few problems i am still working on. For example, the ordering doesnt work as before, since it is not possible to reorder in each facet. I will try to find a solution for this. I also added ggplot2 as an import