major update to abe.resampling
added the option of performing parallel calculations using foreach package.

NOTES: I could not make the example to run, unless I sourced the abe.R (and library(survival) from within the dopar: the workers did not find some functions from abe package or survival package otherwise (although I have specified also .packages option). Note that we needed to add survival in NAMESPACE if we want to make this usable for coxph.

Maybe there is a ways to specify the cluster so that each worker uses the local version of abe and locally installed survival package, but I did not find a way to do it.

I hacked this by adding to lines of code (which are now commented out) and the fitting, as expected, occurs much faster, so in the end this will be worthwile.

NOTE 2: when using dopar using ids generated outside of the dopar loop is a must!